+++ src/TowerInterface.cpp
@@ -7,6 +7,7 @@ TowerInterface::TowerInterface(){
   totalUpgradesPurchased = 0;
   //the id is the address of the current object
   towerID = (long long) this;
+
 }